requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.32.0
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langsmith>=0.1.0
